+++ libs/designsystem/src/lib/components/list/components/list-header/list-header.component.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/utils';
+@import '../../../../scss/utils';
 :host {
   background-color: transparent;
   display: flex;
