+++ libs/designsystem/src/lib/components/list/components/list-section-header/list-section-header.component.ts
