+++ libs/designsystem/src/lib/components/list/list.module.ts
@@ -12,8 +12,8 @@ import {
   ListHeaderDirective,
   ListFooterDirective,
 } from './list.directive';
-import { ListSectionHeaderComponent } from './list-section-header/list-section-header.component';
-import { ListHeaderComponent } from './list-header/list-header.component';
+import { ListSectionHeaderComponent } from './components/list-section-header/list-section-header.component';
+import { ListHeaderComponent } from './components/list-header/list-header.component';
 import { VirtualScrollListComponent } from './components/virtual-scroll-list/virtual-scroll-list.component';
 import { IconModule } from '../icon/icon.module';
 import { SpinnerModule } from '../spinner/spinner.module';
